refactor goto
